refactor: Send response via channel rather than promise
We were using a promise to deliver the http response from nexus to
pedestal. This worked fine, but would have caused the thread to block.

Pedestal allows an interceptor to return a channel and exit without
waiting for the response to arrive. This changeset replaces the promise
with a channel, which should in theory allow for better use of
resources.

This changeset also verifies that the list of actions includes a
:http/response effect before returning the channel, to prevent waiting
on a value from a channel that will never arrive.
When the effect isn't found, the context is returned as is instead,
allowing pedestal's default response handling to kick in.